Add in system for handling database updates
