+++ atlas/core/config.hpp
@@ -306,6 +306,8 @@ SETTINGS_D( experimental, local_match, bool, false )
 SETTINGS_D( experimental, loading_preview, bool, false )
 SETTINGS_D( experimental, loading_preview_blur, int, 0 )
 
+SETTINGS_D( database, migration_version, int, 0 )
+
 #ifdef __GNUC__
 #pragma GCC diagnostic pop
 #endif
