+++ atlas/core/database/Database.cpp
@@ -8,6 +8,7 @@
 
 #include "Transaction.hpp"
 #include "core/config.hpp"
+#include "core/database/migrations/templates.hpp"
 #include "core/database/record/Game.hpp"
 
 namespace internal
@@ -54,51 +55,7 @@ void Database::initalize( const std::filesystem::path init_path )
 
 	RapidTransaction transaction {};
 
-	const std::vector< std::string > table_queries {
-		"CREATE TABLE IF NOT EXISTS games (record_id INTEGER PRIMARY KEY, title TEXT NOT NULL, creator TEXT NOT NULL, engine TEXT, last_played_r DATE DEFAULT 0, total_playtime INTEGER DEFAULT 0, description TEXT, UNIQUE(title, creator, engine));",
-		"CREATE TABLE IF NOT EXISTS versions (record_id INTEGER REFERENCES games(record_id), version TEXT, game_path TEXT, exec_path TEXT, in_place, last_played DATE, version_playtime INTEGER, folder_size INTEGER, date_added INTEGER, UNIQUE(record_id, version));",
-		"CREATE VIEW IF NOT EXISTS last_import_times (record_id, last_import) AS SELECT DISTINCT record_id, versions.date_added FROM games NATURAL JOIN versions ORDER BY versions.date_added DESC;",
-
-		//Atlas data tables
-		"CREATE TABLE IF NOT EXISTS atlas_data (atlas_id INTEGER PRIMARY KEY, id_name STRING UNIQUE, short_name STRING,"
-		"title STRING, original_name STRING, category STRING, engine STRING, status STRING, version STRING,"
-		"developer STRING, creator STRING, overview STRING, censored STRING, language STRING, translations STRING,"
-		"genre STRING, tags STRING, voice STRING, os STRING, release_date DATE, length STRING, banner STRING, banner_wide STRING,"
-		"cover STRING, logo STRING, wallpaper STRING, previews STRING, last_db_update STRING);",
-		"CREATE TABLE IF NOT EXISTS atlas_previews (atlas_id INTEGER REFERENCES atlas_data(atlas_id), preview_url STRING NOT NULL, UNIQUE(atlas_id, preview_url))",
-
-		"CREATE TABLE IF NOT EXISTS atlas_mappings (record_id INTEGER REFERENCES games(record_id), atlas_id INTEGER REFERENCES atlas_data(id), UNIQUE(record_id, atlas_id));",
-
-		//F95 data tables
-		"CREATE TABLE IF NOT EXISTS f95_zone_data (f95_id INTEGER UNIQUE PRIMARY KEY, atlas_id INTEGER REFERENCES atlas_data(atlas_id) UNIQUE , banner_url STRING, site_url STRING,"
-		"last_thread_comment STRING, thread_publish_date STRING, last_record_update STRING, views STRING, likes STRING, tags STRING, rating STRING,"
-		"screens STRING, replies STRING);",
-		"CREATE TABLE IF NOT EXISTS f95_zone_screens (f95_id INTEGER REFERENCES f95_zone_data(f95_id), screen_url TEXT NOT NULL)",
-
-		//Update handling
-		"CREATE TABLE IF NOT EXISTS updates (update_time INTEGER PRIMARY KEY, processed_time INTEGER, md5 BLOB);",
-
-		//Tags
-		"CREATE TABLE IF NOT EXISTS tags (tag_id INTEGER PRIMARY KEY, tag TEXT UNIQUE)",
-		"CREATE TABLE IF NOT EXISTS tag_mappings (record_id INTEGER REFERENCES games(record_id), tag_id REFERENCES tags(tag_id), UNIQUE(record_id, tag_id))",
-		"CREATE TABLE IF NOT EXISTS atlas_tags (tag_id INTEGER REFERENCES tags(tag_id), atlas_id INTEGER REFERENCES atlas_data(atlas_id), UNIQUE(atlas_id, tag_id))",
-		"CREATE TABLE IF NOT EXISTS f95_zone_tags (f95_id INTEGER REFERENCES f95_zone_data(f95_id), tag_id INTEGER REFERENCES tags(tag_id))",
-
-		//Tag views
-		"CREATE VIEW IF NOT EXISTS title_tags (tag, record_id) AS SELECT 'title:' || title, record_id FROM games;",
-		"CREATE VIEW IF NOT EXISTS creator_tags (tag, record_id) AS SELECT 'creator:' || creator, record_id FROM games;",
-		"CREATE VIEW IF NOT EXISTS engine_tags (tag, record_id) AS SELECT 'engine:' || engine, record_id FROM games;",
-		"CREATE VIEW IF NOT EXISTS full_tags (tag, record_id) AS SELECT tag, record_id FROM tags NATURAL JOIN tag_mappings NATURAL JOIN games UNION SELECT tag, record_id FROM title_tags UNION SELECT tag, record_id FROM creator_tags UNION SELECT tag, record_id FROM engine_tags;",
-
-		//Image tables
-		"CREATE TABLE IF NOT EXISTS previews (record_id REFERENCES games(record_id), path TEXT UNIQUE, position INTEGER DEFAULT 256, UNIQUE(record_id, path))",
-		"CREATE TABLE IF NOT EXISTS banners (record_id REFERENCES games(record_id), path TEXT UNIQUE, type INTEGER, UNIQUE(record_id, path, type))",
-
-		//Stats tables
-		"CREATE TABLE IF NOT EXISTS data_change (timestamp INTEGER, delta INTEGER)",
-	};
-
-	for ( const auto& query_str : table_queries ) transaction << query_str;
+	atlas::database::migrations::runUp();
 
 	//Add dummy record
 	const QString test_name { "Galaxy Crossing: First Conquest" };
@@ -109,7 +66,11 @@ void Database::initalize( const std::filesystem::path init_path )
 	{
 		atlas::records::Game game { atlas::records::importRecord( test_name, test_creator, test_engine ) };
 		game.addVersion(
-			"Chapter: 1", "C:/Atlas Games/Galaxy Crossing First Conquest", "Galaxy Crossing First Conquest.exe", 0, true );
+			"Chapter: 1",
+			"C:/Atlas Games/Galaxy Crossing First Conquest",
+			"Galaxy Crossing First Conquest.exe",
+			0,
+			true );
 
 		QImage banner_image { ":/images/assets/Grid_Capsule_Default.webp" };
 		if ( banner_image.isNull() ) throw std::runtime_error( "Failed to open image asset for Grid Capsule Default" );
