+++ atlas/core/database/Transaction.cpp
@@ -4,44 +4,48 @@
 
 #include "Transaction.hpp"
 
-inline static std::atomic< std::thread::id > last_locked {};
-inline static std::mutex trans_lock {};
-
-std::lock_guard< std::mutex > getLock()
+namespace atlas::database
 {
-	if ( last_locked == std::this_thread::get_id() )
-	{
-		//Deadlock
-		throw std::runtime_error( "Deadlock detected!" );
-	}
-	else
-	{
-		last_locked = std::this_thread::get_id();
-		Binder( "BEGIN TRANSACTION" );
-		return std::lock_guard( trans_lock );
-	}
-}
 
-template <>
-TransactionBase< true >::TransactionBase() : guard( getLock() )
-{}
+	inline static std::atomic< std::thread::id > last_locked {};
+	inline static std::mutex trans_lock {};
 
-template <>
-TransactionBase< true >::~TransactionBase() noexcept( false )
-{
-	if ( !m_finished )
+	std::lock_guard< std::mutex > getLock()
 	{
-		Binder( "ROLLBACK TRANSACTION" );
-		throw std::runtime_error( "Allowed falloff via dtor in TransactionBase<true>!. Rolling back and failing." );
+		if ( last_locked == std::this_thread::get_id() )
+		{
+			//Deadlock
+			throw std::runtime_error( "Deadlock detected!" );
+		}
+		else
+		{
+			last_locked = std::this_thread::get_id();
+			Binder( "BEGIN TRANSACTION" );
+			return std::lock_guard( trans_lock );
+		}
 	}
 
-	last_locked = std::thread::id();
-}
+	template <>
+	TransactionBase< true >::TransactionBase() : guard( getLock() )
+	{}
 
-template <>
-TransactionBase< false >::TransactionBase() : guard( self_mtx )
-{}
+	template <>
+	TransactionBase< true >::~TransactionBase() noexcept( false )
+	{
+		if ( !m_finished )
+		{
+			Binder( "ROLLBACK TRANSACTION" );
+			throw std::runtime_error( "Allowed falloff via dtor in TransactionBase<true>!. Rolling back and failing." );
+		}
 
-template <>
-TransactionBase< false >::~TransactionBase() noexcept( false )
-{}
+		last_locked = std::thread::id();
+	}
+
+	template <>
+	TransactionBase< false >::TransactionBase() : guard( self_mtx )
+	{}
+
+	template <>
+	TransactionBase< false >::~TransactionBase() noexcept( false )
+	{}
+} // namespace atlas::database
