+++ atlas/core/logging.hpp
@@ -30,10 +30,6 @@
 	#endif
 #endif // __GNUC__
 
-#include <QJsonArray>
-#include <QJsonDocument>
-#include <QJsonObject>
-
 #include <spdlog/spdlog.h>
 #include <tracy/Tracy.hpp>
 
@@ -100,76 +96,7 @@ namespace atlas::logging
 	inline void userwarn( [[maybe_unused]] std::string message )
 	{
 		spdlog::warn( "userwarn: {}", message );
-		qDebug() << "userwarn: STUB";
 	}
-
-	//! Dev namespace is used for allowing data serialization in order to have the user be able to copy application state to the clipboard and send.
-	namespace dev
-	{
-		//To implement for each class
-		template < typename T >
-		QJsonObject internalSerializer( const T& t );
-
-		//! Injects information from T into the json
-		template < typename T >
-		inline void injectData( [[maybe_unused]] const T& t, [[maybe_unused]] QJsonObject& obj )
-		{
-			//const auto& typeinfo { typeid( T ) };
-			//obj[ "type" ] = QString::fromUtf8( typeinfo.name() );
-		}
-
-		template < typename T >
-		inline QJsonObject serialize( const T& t )
-		{
-			QJsonObject obj { internalSerializer< T >( t ) };
-			injectData< T >( t, obj );
-			return obj;
-		}
-
-		template < typename T, typename... Ts >
-		inline void serializeInto( QJsonArray& array, const T& t, const Ts&... left )
-		{
-			ZoneScoped;
-			const auto data { serialize< T >( t ) };
-			injectData( t, data ); // Insert things like type and pointer information into the object.
-
-			array.push_back( std::move( data ) );
-
-			if constexpr ( sizeof...( left ) > 1 )
-				serializeInto( array, t, left... );
-			else if ( sizeof...( left ) == 1 )
-			{
-				const auto data2 { serialize( left... ) };
-				injectData( left..., data2 );
-				array.push_back( std::move( data2 ) );
-			}
-		}
-
-		template < typename... Ts >
-		inline QJsonDocument serializeObjects( const Ts&... items )
-		{
-			QJsonArray array;
-
-			serializeInto( array, items... );
-
-			QJsonDocument doc;
-			doc.setArray( std::move( array ) );
-			return doc;
-		}
-
-		template < typename... Ts >
-		inline void error( std::string msg, const Ts&... ts )
-		{
-			spdlog::error( "{}: {}", msg, serializeObjects( ts... ).toJson().toStdString() );
-		}
-
-		template <>
-		inline void error<>( std::string msg )
-		{
-			spdlog::error( "{}", msg );
-		}
-
-	} // namespace dev
 
 } // namespace atlas::logging
