+++ atlas/core/notifications/notifications.hpp
@@ -9,6 +9,7 @@
 #include <QString>
 
 #include "core/logging.hpp"
+#include "core/logging/dev.hpp"
 #include "ui/notifications/DevNotification.hpp"
 #include "ui/notifications/GroupedProgressNotification.hpp"
 #include "ui/notifications/MessageNotification.hpp"
